Allow float values for the position in the response to a SetMyDirection command

+++ game/network.py
@@ -126,8 +126,8 @@ class SetMyDirection(Command):
 
     arguments = [('direction', Direction())]
 
-    response = [('x', Integer()),
-                ('y', Integer())]
+    response = [('x', Float()),
+                ('y', Float())]
